change:  - fix cli options  - fix test duration  - add test duration option to cli
git-svn-id: http://mobicents.googlecode.com/svn/trunk@7505 bf0df8d0-2c1f-0410-b170-bd30377b63dc

+++ servers/media/test-suite/core/src/main/java/org/mobicents/media/server/testsuite/general/CallDisplayInterface.java
@@ -16,7 +16,7 @@ import javax.sdp.Attribute;
 public interface CallDisplayInterface {
 
     public static final int _DEFAULT_CPS = 1;
-    public static final int _DEFAULT_CALL_DURATION = 2000;
+    public static final long _DEFAULT_CALL_DURATION = 2000;
     
     //Some methods to get requried fields.
     /**
@@ -31,7 +31,7 @@ public interface CallDisplayInterface {
     public String getRemoteAddress();
     public int getLocalPort();
     public int getRemotePort();
-    public int getCallDuration();
+    public long getCallDuration();
     public int getCPS();
     public int getMaxConcurrentCalls();
     public long getMaxCalls();
